separate character and stats
